api/requirements.txt
# Azure Functions Python dependencies
azure-functions>=1.17.0
openai>=1.12.0
azure-search-documents>=11.4.0

# Semantic Kernel for agent orchestration
semantic-kernel[azure]>=1.27.0

# Tavily for web search
tavily-python>=0.5.0

# Date/time handling
python-dateutil>=2.8.0
pytz>=2024.1
